loudness: merge groveloudness, mic, and hyld9767 into a single driver
