+++ examples/c++/CMakeLists.txt
@@ -29,7 +29,6 @@ add_executable (max31855-example max31855.cxx)
 add_executable (bmpx8x-example bmpx8x.cxx)
 add_executable (stepmotor-example stepmotor.cxx)
 add_executable (pulsensor-example pulsensor.cxx)
-add_executable (mic-example mic.cxx)
 add_executable (mpu9150-example mpu9150.cxx)
 add_executable (maxds3231m-example maxds3231m.cxx)
 add_executable (max31723-example max31723.cxx)
@@ -68,7 +67,7 @@ add_executable (grovelinefinder-example grovelinefinder.cxx)
 add_executable (grovevdiv-example grovevdiv.cxx)
 add_executable (grovewater-example grovewater.cxx)
 add_executable (guvas12d-example guvas12d.cxx)
-add_executable (groveloudness-example groveloudness.cxx)
+add_executable (loudness-example loudness.cxx)
 add_executable (mpr121-example mpr121.cxx)
 add_executable (ublox6-example ublox6.cxx)
 add_executable (yg1006-example yg1006.cxx)
@@ -135,7 +134,6 @@ add_executable (ak8975-example ak8975.cxx)
 add_executable (lsm9ds0-example lsm9ds0.cxx)
 add_executable (eboled-example eboled.cxx)
 add_executable (mpu9250-example mpu9250.cxx)
-add_executable (hyld9767-example hyld9767.cxx)
 add_executable (mg811-example mg811.cxx)
 add_executable (wheelencoder-example wheelencoder.cxx)
 add_executable (sm130-example sm130.cxx)
@@ -156,7 +154,6 @@ include_directories (${PROJECT_SOURCE_DIR}/src/max31855)
 include_directories (${PROJECT_SOURCE_DIR}/src/bmpx8x)
 include_directories (${PROJECT_SOURCE_DIR}/src/stepmotor)
 include_directories (${PROJECT_SOURCE_DIR}/src/pulsensor)
-include_directories (${PROJECT_SOURCE_DIR}/src/mic)
 include_directories (${PROJECT_SOURCE_DIR}/src/mpu9150)
 include_directories (${PROJECT_SOURCE_DIR}/src/maxds3231m)
 include_directories (${PROJECT_SOURCE_DIR}/src/max31723)
@@ -190,7 +187,7 @@ include_directories (${PROJECT_SOURCE_DIR}/src/grovelinefinder)
 include_directories (${PROJECT_SOURCE_DIR}/src/grovevdiv)
 include_directories (${PROJECT_SOURCE_DIR}/src/grovewater)
 include_directories (${PROJECT_SOURCE_DIR}/src/guvas12d)
-include_directories (${PROJECT_SOURCE_DIR}/src/groveloudness)
+include_directories (${PROJECT_SOURCE_DIR}/src/loudness)
 include_directories (${PROJECT_SOURCE_DIR}/src/mpr121)
 include_directories (${PROJECT_SOURCE_DIR}/src/ublox6)
 include_directories (${PROJECT_SOURCE_DIR}/src/yg1006)
@@ -246,7 +243,6 @@ include_directories (${PROJECT_SOURCE_DIR}/src/rgbringcoder)
 include_directories (${PROJECT_SOURCE_DIR}/src/hp20x)
 include_directories (${PROJECT_SOURCE_DIR}/src/pn532)
 include_directories (${PROJECT_SOURCE_DIR}/src/lsm9ds0)
-include_directories (${PROJECT_SOURCE_DIR}/src/hyld9767)
 include_directories (${PROJECT_SOURCE_DIR}/src/mg811)
 include_directories (${PROJECT_SOURCE_DIR}/src/wheelencoder)
 include_directories (${PROJECT_SOURCE_DIR}/src/sm130)
@@ -281,7 +277,6 @@ target_link_libraries (max31855-example max31855 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (bmpx8x-example bmpx8x ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (stepmotor-example stepmotor ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (pulsensor-example pulsensor ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries (mic-example mic ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mpu9150-example mpu9150 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (maxds3231m-example maxds3231m ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (max31723-example max31723 ${CMAKE_THREAD_LIBS_INIT})
@@ -320,7 +315,7 @@ target_link_libraries (grovelinefinder-example grovelinefinder ${CMAKE_THREAD_LI
 target_link_libraries (grovevdiv-example grovevdiv ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (grovewater-example grovewater ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (guvas12d-example guvas12d ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries (groveloudness-example groveloudness ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries (loudness-example loudness ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mpr121-example mpr121 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (ublox6-example ublox6 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (yg1006-example yg1006 ${CMAKE_THREAD_LIBS_INIT})
@@ -387,7 +382,6 @@ target_link_libraries (ak8975-example mpu9150 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (lsm9ds0-example lsm9ds0 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (eboled-example i2clcd ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mpu9250-example mpu9150 ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries (hyld9767-example hyld9767 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mg811-example mg811 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (wheelencoder-example wheelencoder ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (sm130-example sm130 ${CMAKE_THREAD_LIBS_INIT})
