--- examples/c++/hyld9767.cxx
@@ -1,68 +1,0 @@
-/*
- * Author: Jon Trulson <jtrulson@ics.com>
- * Copyright (c) 2015 Intel Corporation.
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-#include <unistd.h>
-#include <iostream>
-#include <signal.h>
-#include "hyld9767.h"
-
-using namespace std;
-
-bool shouldRun = true;
-
-#define HYLD9767_AREF   5.0
-
-void sig_handler(int signo)
-{
-  if (signo == SIGINT)
-    shouldRun = false;
-}
-
-int main()
-{
-  signal(SIGINT, sig_handler);
-
-//! [Interesting]
-
-  // Instantiate a HYLD9767 on analog pin A0, with an analog
-  // reference voltage of HYLD9767_AREF
-  upm::HYLD9767 *loud = new upm::HYLD9767(0, HYLD9767_AREF);
-  
-  // Every tenth of a second, sample the loudness and output it's
-  // corresponding analog voltage. 
-
-  while (shouldRun)
-    {
-      cout << "Detected loudness (volts): " << loud->loudness() << endl;
-      
-      usleep(100000);
-    }
-
-//! [Interesting]
-
-  cout << "Exiting" << endl;
-
-  delete loud;
-  return 0;
-}
