--- src/hyld9767/pyupm_hyld9767.i
@@ -1,9 +1,0 @@
-%module pyupm_hyld9767
-%include "../upm.i"
-
-%feature("autodoc", "3");
-
-%include "hyld9767.h"
-%{
-    #include "hyld9767.h"
-%}
