+++ src/loudness/CMakeLists.txt
@@ -1,5 +1,5 @@
-set (libname "hyld9767")
-set (libdescription "upm DFRobot loudness sensor")
+set (libname "loudness")
+set (libdescription "upm loudness sensors module")
 set (module_src ${libname}.cxx)
 set (module_h ${libname}.h)
 upm_module_init()
