--- src/mic/CMakeLists.txt
@@ -1,5 +1,0 @@
-set (libname "mic")
-set (libdescription "Microphone simple API")
-set (module_src ${libname}.cxx)
-set (module_h ${libname}.h)
-upm_module_init()
